requirements.txt
anthropic
flask
python-dotenv
pillow
openpyxl
pdfplumber
pdf2image
google-cloud-vision